fix(search): Fix hardcoded path to search index
The path to `search.json` was previously hardcoded, causing search
to fail when the site's `baseurl` was changed.

This commit passes the `site.baseurl` value from Jekyll to a global
JavaScript variable, `window.siteBaseUrl`. The search script now
uses this variable to dynamically construct the correct URL for the
search index.